+++ stubs/DecodeLabs/Harvest.php
@@ -8,26 +8,28 @@ namespace DecodeLabs;
 use DecodeLabs\Veneer\Proxy as Proxy;
 use DecodeLabs\Veneer\ProxyTrait as ProxyTrait;
 use DecodeLabs\Harvest\Context as Inst;
-use DecodeLabs\Singularity\Url as Ref0;
-use Psr\Http\Message\StreamInterface as Ref1;
-use DecodeLabs\Atlas\File as Ref2;
-use DecodeLabs\Harvest\Request as Ref3;
-use DecodeLabs\Harvest\Transport as Ref4;
-use Psr\Http\Message\ServerRequestInterface as Ref5;
-use Psr\Http\Message\ResponseInterface as Ref6;
-use DecodeLabs\Deliverance\Channel\Stream as Ref7;
-use DecodeLabs\Harvest\Response\Stream as Ref8;
-use Stringable as Ref9;
-use Generator as Ref10;
-use Closure as Ref11;
-use DecodeLabs\Harvest\Response\Text as Ref12;
-use DecodeLabs\Harvest\Response\Html as Ref13;
-use DecodeLabs\Harvest\Response\Json as Ref14;
-use DecodeLabs\Harvest\Response\Xml as Ref15;
-use Psr\Http\Message\UriInterface as Ref16;
-use DecodeLabs\Harvest\Response\Redirect as Ref17;
-use Traversable as Ref18;
-use DecodeLabs\Compass\Ip as Ref19;
+use DecodeLabs\Harvest\Cookie\Collection as CookiesPlugin;
+use DecodeLabs\Harvest\Profile as Ref0;
+use DecodeLabs\Singularity\Url as Ref1;
+use Psr\Http\Message\StreamInterface as Ref2;
+use DecodeLabs\Atlas\File as Ref3;
+use DecodeLabs\Harvest\Request as Ref4;
+use DecodeLabs\Harvest\Transport as Ref5;
+use Psr\Http\Message\ServerRequestInterface as Ref6;
+use Psr\Http\Message\ResponseInterface as Ref7;
+use DecodeLabs\Deliverance\Channel\Stream as Ref8;
+use DecodeLabs\Harvest\Response\Stream as Ref9;
+use Stringable as Ref10;
+use Generator as Ref11;
+use Closure as Ref12;
+use DecodeLabs\Harvest\Response\Text as Ref13;
+use DecodeLabs\Harvest\Response\Html as Ref14;
+use DecodeLabs\Harvest\Response\Json as Ref15;
+use DecodeLabs\Harvest\Response\Xml as Ref16;
+use Psr\Http\Message\UriInterface as Ref17;
+use DecodeLabs\Harvest\Response\Redirect as Ref18;
+use Traversable as Ref19;
+use DecodeLabs\Compass\Ip as Ref20;
 
 class Harvest implements Proxy
 {
@@ -37,53 +39,57 @@ class Harvest implements Proxy
     public const VeneerTarget = Inst::class;
 
     protected static Inst $_veneerInstance;
+    public static CookiesPlugin $cookies;
 
-    public static function createUri(string $uri = ''): Ref0 {
+    public static function loadDefaultProfile(): Ref0 {
+        return static::$_veneerInstance->loadDefaultProfile();
+    }
+    public static function createUri(string $uri = ''): Ref1 {
         return static::$_veneerInstance->createUri(...func_get_args());
     }
-    public static function createStream(string $content = ''): Ref1 {
+    public static function createStream(string $content = ''): Ref2 {
         return static::$_veneerInstance->createStream(...func_get_args());
     }
-    public static function createStreamFromFile(Ref2|string $filename, string $mode = 'r'): Ref1 {
+    public static function createStreamFromFile(Ref3|string $filename, string $mode = 'r'): Ref2 {
         return static::$_veneerInstance->createStreamFromFile(...func_get_args());
     }
-    public static function createStreamFromResource($resource): Ref1 {
+    public static function createStreamFromResource($resource): Ref2 {
         return static::$_veneerInstance->createStreamFromResource(...func_get_args());
     }
-    public static function createRequestFromEnvironment(?string $method = NULL, $uri = NULL, ?array $server = NULL): Ref3 {
+    public static function createRequestFromEnvironment(?string $method = NULL, $uri = NULL, ?array $server = NULL): Ref4 {
         return static::$_veneerInstance->createRequestFromEnvironment(...func_get_args());
     }
-    public static function createTransport(?string $name = NULL): Ref4 {
+    public static function createTransport(?string $name = NULL): Ref5 {
         return static::$_veneerInstance->createTransport(...func_get_args());
     }
-    public static function transform(Ref5 $request, mixed $response): Ref6 {
+    public static function transform(Ref6 $request, mixed $response): Ref7 {
         return static::$_veneerInstance->transform(...func_get_args());
     }
-    public static function stream(Ref7|Ref1|string $body = 'php://memory', int $status = 200, array $headers = []): Ref8 {
+    public static function stream(Ref8|Ref2|string $body = 'php://memory', int $status = 200, array $headers = []): Ref9 {
         return static::$_veneerInstance->stream(...func_get_args());
     }
-    public static function text(Ref9|Ref10|Ref11|string $text, int $status = 200, array $headers = []): Ref12 {
+    public static function text(Ref10|Ref11|Ref12|string $text, int $status = 200, array $headers = []): Ref13 {
         return static::$_veneerInstance->text(...func_get_args());
     }
-    public static function html(Ref9|Ref10|Ref11|string $html, int $status = 200, array $headers = []): Ref13 {
+    public static function html(Ref10|Ref11|Ref12|string $html, int $status = 200, array $headers = []): Ref14 {
         return static::$_veneerInstance->html(...func_get_args());
     }
-    public static function json(mixed $data, int $status = 200, array $headers = []): Ref14 {
+    public static function json(mixed $data, int $status = 200, array $headers = []): Ref15 {
         return static::$_veneerInstance->json(...func_get_args());
     }
-    public static function xml(Ref9|Ref10|Ref11|string $xml, int $status = 200, array $headers = []): Ref15 {
+    public static function xml(Ref10|Ref11|Ref12|string $xml, int $status = 200, array $headers = []): Ref16 {
         return static::$_veneerInstance->xml(...func_get_args());
     }
-    public static function redirect(Ref16|string $uri, int $status = 302, array $headers = []): Ref17 {
+    public static function redirect(Ref17|string $uri, int $status = 302, array $headers = []): Ref18 {
         return static::$_veneerInstance->redirect(...func_get_args());
     }
-    public static function generator(Ref18|Ref11|array $iterator, int $status = 200, array $headers = []): Ref8 {
+    public static function generator(Ref19|Ref12|array $iterator, int $status = 200, array $headers = []): Ref9 {
         return static::$_veneerInstance->generator(...func_get_args());
     }
-    public static function liveGenerator(Ref18|Ref11|array $iterator, int $status = 200, array $headers = []): Ref8 {
+    public static function liveGenerator(Ref19|Ref12|array $iterator, int $status = 200, array $headers = []): Ref9 {
         return static::$_veneerInstance->liveGenerator(...func_get_args());
     }
-    public static function extractIpFromRequest(Ref5 $request): Ref19 {
+    public static function extractIpFromRequest(Ref6 $request): Ref20 {
         return static::$_veneerInstance->extractIpFromRequest(...func_get_args());
     }
 };